pyproject.toml
[tool.poetry]
authors = ["Stig B. Dørmænen <stigbd@gmail.com>"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: Apache Software License",
  "Operating System :: OS Independent",
]
description = "The service validates a graph against dcat-ap-no specification"
homepage = "https://github.com/stigbd/aiohttp-hello-world"
license = "Apache-2.0"
name = "aiohttp-hello-world"
readme = "README.md"
repository = "https://github.com/stigbd/aiohttp-hello-world"
version = "0.1.0"

[tool.poetry.dependencies]
aiohttp = "^3.8.1"
aiohttp-middlewares = "^2.0.0"
certifi = "^2022.6.15"
dataclasses-json = "^0.5.7"
gunicorn = "^20.1.0"
motor = "^3.0.0"
python = ">=3.7,<3.11"
python-dotenv = "^0.20.0"
python-json-logger = "^2.0.2"

[tool.poetry.dev-dependencies]
aiohttp-devtools = "^1.0.post0"
black = "^22.6.0"
codecov = "^2.1.12"
coverage = "^6.4.1"
docker-compose = "^1.29.2"
flake8 = "^4.0.1"
flake8-annotations = "^2.9.0"
flake8-bandit = "^3.0.0"
flake8-black = "^0.3.3"
flake8-bugbear = "^22.6.22"
flake8-docstrings = "^1.6.0"
flake8-import-order = "^0.18.1"
jsondiff = "^2.0.0"
mypy = "^0.971"
nox = "^2022.1.7"
nox-poetry = "^1.0.1"
pep8-naming = "^0.13.0"
pytest = "^7.1.2"
pytest-aiohttp = "^1.0.4"
pytest-asyncio = "^0.19.0"
pytest-cov = "^3.0.0"
pytest-docker = "^1.0.0"
pytest-mock = "^3.8.1"
pytest-profiling = "^1.7.0"
requests = "^2.28.0"
safety = "^2.1.1"
types-requests = "^2.28.8"

[tool.coverage.paths]
source = ["aiohttp_hello_world", "*/site-packages"]

[tool.coverage.run]
branch = false
omit = ["*gunicorn_config.py"]
source = ["aiohttp_hello_world"]

[tool.coverage.report]
fail_under = 100
show_missing = true

[tool.mypy]
pretty = true
show_column_numbers = true
show_error_codes = true
show_error_context = true
strict = false
warn_unreachable = true

[[tool.mypy.overrides]]
ignore_missing_imports = true
module = [
  "gunicorn.*",
  "jsondiff.*",
  "motor.*",
  "pytest_mock.*",
  "pythonjsonlogger.*",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = [
  "unit: marks tests as unit (fast)",
  "integration: marks tests as integration (slower)",
  "contract: marks test as contract (slow)",
]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]
